+++ util/kstdlib/src/kstdio.c
@@ -5,10 +5,12 @@
 
 #include <stdarg.h>
 
-size_t kprintf(const char *a_format, ...)
+uint32 kprintf(
+    const char *a_format,
+    ...)
 {
-    if (a_format == NULL)
-    {
+    if (a_format == NULL) {
+
         return ERROR_NULL_POINTER;
     }
 
@@ -34,17 +36,17 @@ size_t kprintf(const char *a_format, ...)
     va_list arg;
     va_start(arg, a_format);
 
-    while ((c = *a_format) != '\0')
-    {
+    while ((c = *a_format) != '\0') {
+
         ++a_format;
 
-        if (c == '%' || extend)
-        {
-            switch (*a_format)
-            {
+        if (c == '%' || extend) {
+
+            switch (*a_format) {
+
                 case 'd':
-                    if (extend)
-                    {
+                    if (extend) {
+
                         esvalue = va_arg(arg, sint64);
 
                         bufferSize = 64;
@@ -52,8 +54,8 @@ size_t kprintf(const char *a_format, ...)
 
                         VGA_WriteString(ebuffer);
                     }
-                    else
-                    {
+                    else {
+
                         svalue = va_arg(arg, sint32);
 
                         bufferSize = 32;
@@ -65,8 +67,8 @@ size_t kprintf(const char *a_format, ...)
                     break;
 
                 case 'u':
-                    if (extend)
-                    {
+                    if (extend) {
+
                         euvalue = va_arg(arg, uint64);
 
                         bufferSize = 64;
@@ -74,8 +76,8 @@ size_t kprintf(const char *a_format, ...)
 
                         VGA_WriteString(ebuffer);
                     }
-                    else
-                    {
+                    else {
+
                         uvalue = va_arg(arg, uint32);
 
                         bufferSize = 32;
@@ -89,12 +91,12 @@ size_t kprintf(const char *a_format, ...)
                 case 's':
                     str = va_arg(arg, const char*);
 
-                    if (str == NULL)
-                    {
+                    if (str == NULL) {
+
                         VGA_WriteString("(null)");
                     }
-                    else
-                    {
+                    else {
+
                         VGA_WriteString(str);
                     }
                     extend = false;
@@ -120,8 +122,8 @@ size_t kprintf(const char *a_format, ...)
                     break;
 
                 case 'x':
-                    if (extend)
-                    {
+                    if (extend) {
+
                         eaddress = va_arg(arg, uint64);
 
                         bufferSize = 64;
@@ -134,8 +136,8 @@ size_t kprintf(const char *a_format, ...)
                         VGA_WriteChar('0');
                         VGA_WriteString(ebuffer);
                     }
-                    else
-                    {
+                    else {
+
                         address = va_arg(arg, uint32);
 
                         bufferSize = 32;
@@ -152,8 +154,8 @@ size_t kprintf(const char *a_format, ...)
                     break;
 
                 case 'l':
-                    if (*(a_format + 1) == 'l') 
-                    {
+                    if (*(a_format + 1) == 'l')  {
+
                         extend = true;
                     }
                     break;
@@ -164,8 +166,8 @@ size_t kprintf(const char *a_format, ...)
             }
             ++a_format;
         }
-        else
-        {
+        else {
+
             VGA_WriteChar(c);
         }
     }
